Refactor dataset utilities in notebooks.support
